@@ -8,46 +8,46 @@
     "build": "webpack --env.dev",
     "build:prod": "webpack --env.prod",
     "relay": "relay-compiler --src ./src --schema schema.graphql --language typescript",
-    "start": "webpack-dev-server --env.devServer"
+    "start": "webpack-dev-server"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@types/react-router": "^5.1.4",
-    "@types/react-router-dom": "^5.1.3",
+    "@types/react-router": "^5.1.5",
+    "@types/react-router-dom": "^5.1.4",
     "classnames": "^2.2.6",
-    "react": "experimental",
-    "react-dom": "experimental",
-    "react-relay": "experimental",
-    "react-router": "^5.1.2",
-    "react-router-dom": "^5.1.2"
+    "react": "0.0.0-experimental-e5d06e34b",
+    "react-dom": "0.0.0-experimental-e5d06e34b",
+    "react-relay": "0.0.0-experimental-8cc94ddc",
+    "react-router": "^6.0.0-alpha.3",
+    "react-router-dom": "^6.0.0-alpha.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.7.7",
-    "@babel/preset-env": "^7.7.7",
-    "@babel/preset-react": "^7.7.4",
-    "@babel/preset-typescript": "^7.7.7",
-    "@types/classnames": "^2.2.9",
-    "@types/express": "^4.17.2",
-    "@types/node": "^13.1.6",
-    "@types/react-dom": "^16.9.4",
+    "@babel/core": "^7.9.0",
+    "@babel/preset-env": "^7.9.5",
+    "@babel/preset-react": "^7.9.4",
+    "@babel/preset-typescript": "^7.9.0",
+    "@types/classnames": "^2.2.10",
+    "@types/express": "^4.17.6",
+    "@types/node": "^13.13.1",
+    "@types/react-dom": "^16.9.6",
     "@types/react-relay": "^7.0.3",
     "@types/relay-runtime": "^6.0.11",
-    "babel-loader": "^8.0.6",
+    "babel-loader": "^8.1.0",
     "babel-plugin-relay": "^7.1.0",
     "clean-webpack-plugin": "^3.0.0",
-    "css-loader": "^3.4.2",
+    "css-loader": "^3.5.2",
     "file-loader": "^4.3.0",
-    "graphql": "^14.5.8",
-    "html-webpack-plugin": "^4.0.0-beta.11",
+    "graphql": "^14.6.0",
+    "html-webpack-plugin": "^4.2.0",
     "mini-css-extract-plugin": "^0.9.0",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
     "preload-webpack-plugin": "^2.3.0",
     "relay-compiler": "^8.0.0",
-    "relay-compiler-language-typescript": "^10.1.4",
-    "typescript": "^3.7.4",
-    "webpack": "^4.41.5",
-    "webpack-cli": "^3.3.10",
-    "webpack-dev-server": "^3.10.1"
+    "relay-compiler-language-typescript": "^10.1.5",
+    "typescript": "^3.8.3",
+    "webpack": "^4.42.1",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3"
   }
 }
